+++ Samples/2.0/Tutorials/Tutorial02_VariableFramerate/Tutorial02_VariableFramerate.cpp
@@ -79,7 +79,7 @@ int mainApp( int argc, const char *argv[] )
     graphicsSystem.createScene02();
 
     Ogre::Timer timer;
-    unsigned long startTime = timer.getMicroseconds();
+    unsigned long long startTime = timer.getMicroseconds();
 
     double timeSinceLast = 1.0 / 60.0;
 
@@ -96,7 +96,7 @@ int mainApp( int argc, const char *argv[] )
             Ogre::Threads::Sleep( 500 );
         }
 
-        unsigned long endTime = timer.getMicroseconds();
+        unsigned long long endTime = timer.getMicroseconds();
         timeSinceLast = (endTime - startTime) / 1000000.0;
         timeSinceLast = std::min( 1.0, timeSinceLast ); //Prevent from going haywire.
         startTime = endTime;
