+++ Samples/2.0/Tutorials/Tutorial04_InterpolationLoop/Tutorial04_InterpolationLoop.cpp
@@ -82,7 +82,7 @@ int mainApp( int argc, const char *argv[] )
 
     Ogre::Timer timer;
 
-    unsigned long startTime = timer.getMicroseconds();
+    unsigned long long startTime = timer.getMicroseconds();
     double accumulator = cFrametime;
 
     double timeSinceLast = 1.0 / 60.0;
@@ -121,7 +121,7 @@ int mainApp( int argc, const char *argv[] )
         if( gFakeFrameskip )
             Ogre::Threads::Sleep( 40 );
 
-        unsigned long endTime = timer.getMicroseconds();
+        unsigned long long endTime = timer.getMicroseconds();
         timeSinceLast = (endTime - startTime) / 1000000.0;
         timeSinceLast = std::min( 1.0, timeSinceLast ); //Prevent from going haywire.
         accumulator += timeSinceLast;
